Clean.Types: Simplified ::Type; implemented Martelli-Montanari
This fixes bugs in previous versions of unify for Types. The whole
function was rewritten. Currently, it is an implementation of Algorithm
1 from section 2 of An Efficient Unification Algorithm, Martelli and
Montanari, 1982.
http://moscova.inria.fr/~levy/courses/X/IF/03/pi/levy2/martelli-montanari.pdf

Also, ::Type was simplified. Lists, Tuples and Arrays are actually
instances of Type String [Type] and were removed. For the moment, we do
not have to consider strictness and other type attributes.