Fixed Data.Array's takeArr
The inner function "copyArr" would never copy the last element of the array if its size > 1.
If the function reaches i == sz - 1 (last element of the array), it should return the new array with the last element copied. Instead, it just returned the new array.
An easy fix is to return the new array only when i == sz, which means that you've copied the entire array alread (including the last element)