Fixing termios flags on macOS
Accordinng to http://kirste.userpage.fu-berlin.de/chemnet/use/info/libc/libc_12.html:
"CLOCAL
If this bit is set, it indicates that the terminal is connected "locally" and that the modem status lines (such as carrier detect) should be ignored.
On many systems if this bit is not set and you call open without the O_NONBLOCK flag set, open blocks until a modem connection is established."

Apparently (by testing and from a tip online), macOS is one of those systems.